requirements.txt
flask
flask_restful
flask_script
flask_migrate
marshmallow
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
passlib
flask-jwt-extended
flask-mail
